requirements.txt
anthropic
python-dotenv
requests
schedule
